[JENKINS-38978] Update to Jenkins 2
The plugin parent was updated to the latest version _3.2_. This will
finally remove support for Jenkins 1.x - which now is quite old enough
not to be supported anymore - and thus implies bumping to a new major
version of this plugin.

The Jenkins version was set to _2.32.1_ so that the plugin can be
updated to the new version even on older Jenkins installations (up to
about a year of age).

The new parent POM requires (or at least encourage) seveal changes to
the project POM:
* The Java level was set to _8_.
* The parent POM enforces the Maven version, hence ``prerequisites``
  is obsolete.
* ``description`` and ``inceptionYear`` from the parent POM should be
  overridden.
* The ``url`` ws applied to _jenkins.io_.
* Some properties and plugin configurations have been removed, because
  they are already configured in the parent POM.