Update credentials plugin and some dependencies
[Benjamin Bonnet](https://github.com/bonnetb) provided a pipeline test
job - many thanks for that! In order to get it merged (cherry-picked)
and running, the credentials plugin needs to be updated. It was updated
to version _2.1.10_ to provide backward compatibility for about a year.
Furthermore, some transitive (test) dependencies need to be explicitly
defined to satisfy the upper bound requirement for plugin dependencies.